refactor : 도메인 필드명 및 타입 수정

+++ src/main/java/com/example/coinProject/coin/repository/CoinRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface CoinRepository extends JpaRepository<Coin, Integer> {
+public interface CoinRepository extends JpaRepository<Coin, Long> {
 
     Optional<CoinResponse> findCoinByMarket (String market);
     Coin findFirstByMarket(String market);
